requirements-phase1.txt
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
colorama==0.4.6
cryptography==46.0.5
faiss-cpu==1.9.0.post1
filelock==3.25.2
fsspec==2026.2.0
huggingface_hub==0.36.2
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
packaging==26.0
pdfminer.six==20221105
pillow==12.1.1
pluggy==1.6.0
pycparser==3.0
pypdf==4.3.1
pytest==8.2.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==3.3.1
setuptools==82.0.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.66.4
transformers==4.57.6
typing_extensions==4.15.0
urllib3==2.6.3
